Unify bootstrap/verify scripts via wrappers to in-tree scripts
